Refactor custom pagy navigation and add CSS classes
The update refactored the way the pagy navigation is generated in the `pagy_nav_custom` method. It now explicitly includes 'previous' and 'next' links, removing these as cases inside the iteration over pagy.series. The change also introduces custom CSS classes for styling the pagination links. Furthermore, CSS styling was added for the pagy navigation block in the Organismes index view. Finally, a new initial Pagy configuration was added to set the items_param to :page.