Update sorting mechanism in organisms_controller and add export functionality in export_to_excel
The update modifies the organism sorting by name in organisme_controller, making it easier to locate specific organisms. In addition, implemented a functionality to export organism data to Excel in the 'export_to_excel.xlsx.axlsx' view, thus improving data portability and ease of use. This also includes relevant headers for the output Excel file and adjustments to the Dockerfile for Ruby version update.