Enhance organism filtering system and add data export functionality
The changes improve the organism searching and filtering functionalities in the Organismes index view, by leveraging the ransack gem. They also include an export to Excel feature in the Organismes index view. Thanks to these modifications, users can now narrow down the list of organisms based on multiple parameters and export the filtered data to an Excel file for further analysis. Paging settings in the Pagy initializer are also updated to enhance user experience in navigating through the list of organisms.