Refine organismes index view and enhance filtering
The index view for 'organismes' was overhauled to provide a more user-friendly and informative experience, with improvements to the badge classification system and the statuses shown. The filtering and sorting capabilities in the organisms controller have also been enhanced by including 'bureau' and 'controller' in the query results. These changes aim to streamline data presentation and offer a more intuitive navigation experience.