Refine organismes controller and update operator label
Updated the organismes controller search feature and enhanced the functionality for resetting values. The operator label in the organismes index view was also updated to display a specific year instead of the current year. Also, added a condition to only display the 'Add a new document' button to specific users.