Implement universal search functionality for Organisme model
Added a universal search functionality to the Organisme model that allows users to search by "nom" or "acronyme". Updated the search form in the header and the organismes index to reflect these changes. Enhanced the design and behavior of the search input field and search results pop-up through CSS and Javascript. Also, introduced a new method in Organismes controller to handle the new search requests.