Add control document deletion functionality and flash notification
Implemented the ability to delete control documents through an 'edit' action in the ControlDocumentsController. Additionally, flash messages have been added to the 'control_documents' views to notify the user when a document has been successfully deleted. Adjustments were made in several views for clarity and better interaction, including the addition of a modal for confirming document deletion.