Update terms in organismes index view
The wording of terms in the organismes index view has been changed to provide clearer and more consistent language. This includes changes to categories such as 'Nature du contrôle', 'Autorité de contrôle', and others in the filter section. This update will improve the user experience and understandability of the application.